:recycle: :: repository 레이어 구조 정리 및 상태 분리
