+++ lib/core/ui/dialog/laundry_status_dialog.dart
@@ -11,6 +11,7 @@ import 'package:washer/core/utils/date_time_formatter.dart';
 import 'package:washer/core/utils/room_formatter.dart';
 import 'package:washer/features/reservation/data/models/local/active_reservation_model.dart';
 import 'package:washer/features/reservation/presentation/providers/reservation_status_provider.dart';
+import 'package:washer/features/reservation/presentation/states/reservation_action_state.dart';
 import 'package:washer/features/reservation/presentation/viewmodels/reservation_view_model.dart';
 
 class LaundryStatusDialog extends ConsumerWidget {
