+++ lib/features/auth/domain/repositories/auth_repository.dart
@@ -1,0 +1,5 @@
+abstract class AuthRepository {
+  Future<void> login(String code);
+  Future<void> logout();
+  Future<void> refresh();
+}
