+++ lib/features/auth/presentation/viewmodels/auth_callback_view_model.dart
@@ -1,7 +1,8 @@
 import 'dart:async';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:washer/features/auth/data/repositories/auth_repository.dart';
+import 'package:washer/features/auth/data/repositories/auth_repository_impl.dart';
+import 'package:washer/features/auth/domain/repositories/auth_repository.dart';
 
 class AuthCallbackViewModel extends AsyncNotifier<void> {
   late final AuthRepository _authRepository;
