+++ lib/features/auth/presentation/viewmodels/logout_view_model.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:washer/core/network/auth_notifier.dart';
-import 'package:washer/features/auth/data/repositories/auth_repository.dart';
+import 'package:washer/features/auth/data/repositories/auth_repository_impl.dart';
+import 'package:washer/features/auth/domain/repositories/auth_repository.dart';
 
 class LogoutViewModel extends AsyncNotifier<void> {
   late final AuthRepository _authRepository;
