+++ lib/features/history/data/repositories/history_repository_impl.dart
@@ -1,16 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:washer/features/history/data/data_sources/history_remote_data_source.dart';
 import 'package:washer/features/history/data/models/machine_history_response.dart';
-
-abstract class HistoryRepository {
-  Future<MachineHistoryResponse> getMachineHistory({
-    required int machineId,
-    required String startDate,
-    required String endDate,
-    int page = 0,
-    int size = 20,
-  });
-}
+import 'package:washer/features/history/domain/repositories/history_repository.dart';
 
 class HistoryRepositoryImpl implements HistoryRepository {
   final HistoryRemoteDataSource _dataSource;
