+++ lib/features/report/data/repositories/report_repository_impl.dart
@@ -1,12 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:washer/features/report/data/data_sources/remote/report_remote_data_source.dart';
-
-abstract class ReportRepository {
-  Future<void> createMalfunctionReport({
-    required int machineId,
-    required String description,
-  });
-}
+import 'package:washer/features/report/domain/repositories/report_repository.dart';
 
 class ReportRepositoryImpl implements ReportRepository {
   final ReportRemoteDataSource _dataSource;
