+++ lib/features/report/domain/repositories/report_repository.dart
@@ -1,0 +1,6 @@
+abstract class ReportRepository {
+  Future<void> createMalfunctionReport({
+    required int machineId,
+    required String description,
+  });
+}
