+++ lib/features/report/presentation/states/report_action_state.dart
@@ -1,0 +1,11 @@
+enum ReportActionStatus { idle, loading, success, error }
+
+class ReportActionState {
+  const ReportActionState({
+    this.status = ReportActionStatus.idle,
+    this.errorMessage,
+  });
+
+  final ReportActionStatus status;
+  final String? errorMessage;
+}
