+++ lib/features/reservation/data/repositories/home_repository_impl.dart
@@ -2,11 +2,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:washer/features/reservation/data/data_sources/remote/reservation_status_remote_data_source.dart';
 import 'package:washer/features/reservation/data/models/local/active_reservation_model.dart';
 import 'package:washer/features/reservation/data/models/local/laundry_machine_model.dart';
-
-abstract class HomeRepository {
-  Future<MachineStatusResponse> getMachineStatus();
-  Future<List<ActiveReservationModel>> getActiveReservations();
-}
+import 'package:washer/features/reservation/domain/repositories/home_repository.dart';
 
 class HomeRepositoryImpl implements HomeRepository {
   final HomeRemoteDataSource _dataSource;
