+++ lib/features/reservation/presentation/states/reservation_action_state.dart
@@ -1,0 +1,15 @@
+import 'package:washer/features/reservation/data/models/local/active_reservation_model.dart';
+
+enum ReservationActionStatus { idle, loading, success, error }
+
+class ReservationActionState {
+  const ReservationActionState({
+    this.status = ReservationActionStatus.idle,
+    this.errorMessage,
+    this.reservation,
+  });
+
+  final ReservationActionStatus status;
+  final String? errorMessage;
+  final ActiveReservationModel? reservation;
+}
