+++ lib/features/reservation/presentation/widgets/reservation_section_widget.dart
@@ -14,6 +14,7 @@ import 'package:washer/core/theme/typography.dart';
 import 'package:washer/features/reservation/data/models/local/active_reservation_model.dart';
 import 'package:washer/features/reservation/data/models/local/laundry_machine_model.dart';
 import 'package:washer/features/reservation/presentation/providers/reservation_status_provider.dart';
+import 'package:washer/features/reservation/presentation/states/reservation_action_state.dart';
 import 'package:washer/features/reservation/presentation/viewmodels/reservation_view_model.dart';
 import 'package:washer/features/user/presentation/viewmodels/my_user_view_model.dart';
 import 'package:washer/features/reservation/presentation/widgets/reservation_widget.dart';
