+++ lib/features/user/data/repositories/user_repository_impl.dart
@@ -1,11 +1,7 @@
-﻿import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:washer/features/user/data/data_sources/remote/user_remote_data_source.dart';
 import 'package:washer/features/user/data/models/my_user_model.dart';
-
-abstract class UserRepository {
-  Future<MyUserModel?> getMyUser();
-  Future<void> withdraw();
-}
+import 'package:washer/features/user/domain/repositories/user_repository.dart';
 
 class UserRepositoryImpl implements UserRepository {
   const UserRepositoryImpl(this._dataSource);
