+++ lib/features/user/domain/repositories/user_repository.dart
@@ -1,0 +1,6 @@
+import 'package:washer/features/user/data/models/my_user_model.dart';
+
+abstract class UserRepository {
+  Future<MyUserModel?> getMyUser();
+  Future<void> withdraw();
+}
